Add encoder with ability to encode strings
Currently only writes small strings into the image by replacing by
iterating through each byte in the image (in reverse), modifying the
least significant bit to be the next bit in the string of data
to encode.

There is an issue with multiple image formats where writing from the
beginning (or data that is too large) will overwrite the header data
and not just the image content, corrupting the image. This will need
to be accommodated in the future.