Remove checking whether output file exists
Confirming overwrite required a stdin input from the user. Since I have
decided to also make this usable as a gem, rather than just a
standalone script, it is important that any class can run this without
printing or requiring user input.

In the future I will be adding a verbose mode which will print
statements as it goes along. This will allow the standalone script
to display messages to the user, while ensure the class can be used in
other projects.