Refactor and add decent documentation for unveiler
Encoder and Decoder classes have been deleted, and the methods have
been refactored into Unveiler. This is to keep the code clean, simple
and intuitive. All file reading and writing has been moved to the bin
folder in order to avoid obstructing use of unveiler for other purposes.

Tests need to be written for the remaining Unveiler methods.