Optimise process_bytes
The algorithm had taken over 30 minutes and still hadn't finished
decoding Othello, so I have optimised it to take around a minute.
This involved changing the algorithm to only compare binary, instead
of converting the string to UTF-8 and checking for a match with EOF
at the end of each byte. This could be further optimised in the future.